refactor composable return type to separate file
